f2

+++ test_files.php
@@ -11,16 +11,16 @@
 <?php
 require 'models'.DIRECTORY_SEPARATOR.'Track.php';
 require 'Utf8.php';
+date_default_timezone_set('Europe/Kiev');
 
 Config::getConfig();
 
 $tracks = Track::find('all'/*, array('conditions' => 'id = 1')*/);
 
 foreach ($tracks as $track) {
-    if(strpos($track->file, '%') !== FALSE){
+    $full_track_path = Track::$search_tracks_in . stripSpaces($track->composer->name) . DIRECTORY_SEPARATOR . stripSpaces($track->name_ascii) . '.mp3';
 
-        $full_track_path = Track::$search_tracks_in . stripSpaces($track->composer->name) . DIRECTORY_SEPARATOR . stripSpaces($track->name_ascii) . '.mp3';
-
+    if(strpos($track->file, '%') !== FALSE && is_file($full_track_path)){
         $stream = file_get_contents($track->file);
         $bytes = file_put_contents($full_track_path, $stream);
         echo "Downloaded #$track->id " . humanFileSize($bytes, 'MB') . " @ " . date("d.m.Y H:i:s" , time()) . "<br>";
